+++ src/components/datatable/Veteriniare.jsx
@@ -11,6 +11,7 @@ import NewVeterinaireVisite from '../../pages/new/NewVeterinaireVisite/NewVeteri
 import { allveterinaire, allvisite, deleteveterinaire, deletevisite, updateveterinaire, updatevisite, value } from '../../var';
 
 const Datatable = () => {
+    let value = 0
 
     const [visite, setVisite] = useState([]);
     const [veterinaire, setVeterinaire] = useState([]);
@@ -41,29 +42,40 @@ const Datatable = () => {
     const handleDeletevisite = (id) => {
 
         console.log('Printing id', id);
-        axios.delete(deletevisite+`${id}`).catch(error => {
-            setSnackbar({ children: error.message, severity: 'error' });
-        })
-        setSnackbar({ children: 'Bien supprimer', severity: 'success' });
-        window.location.reload(false);
+        if (window.confirm('Êtes-vous sûr de vouloir enregistrer cet élément dans la base de données ?')) {
+            axios.delete(deletevisite + `${id}`).catch(error => {
+                setSnackbar({ children: error.message, severity: 'error' });
+            })
+            setSnackbar({ children: 'Bien supprimer', severity: 'success' });
+            window.location.reload(false);
+        } else {
+            console.log('Thing was not saved to the database.');
+        }
+       
 
     }
     const handleDeleteveterinaire = (id) => {
 
         console.log('Printing id', id);
-        axios.delete(deleteveterinaire+`${id}`).catch(error => {
-            if (error) {
-                const timer = setTimeout(() => {
-                    setSnackbar({ children: " Cet veterinaire a déja effectué une viste!", severity: 'error' });
-                }, 500);
+        if (window.confirm('Êtes-vous sûr de vouloir enregistrer cet élément dans la base de données ?')) {
+            axios.delete(deleteveterinaire + `${id}`).catch(error => {
+                if (error) {
+                    const timer = setTimeout(() => {
+                        setSnackbar({ children: " Cet veterinaire a déja effectué une viste!", severity: 'error' });
+                    }, 500);
 
-                return () => clearTimeout(timer);
-            }
+                    return () => clearTimeout(timer);
+                }
 
-            setSnackbar({ children: 'Bien supprimer', severity: 'success' });
+                setSnackbar({ children: 'Bien supprimer', severity: 'success' });
 
-            window.location.reload(false);
-        })
+                window.location.reload(false);
+            })
+        } else {
+            console.log('Thing was not saved to the database.');
+        }
+
+        
 
     }
     useEffect(() => {
@@ -87,7 +99,7 @@ const Datatable = () => {
 
             return (
                 <div className="cellAction">
-                    <div className="deleteButton" >Delete</div>
+                    <div className="deleteButton" >Supprimer</div>
                 </div>
             )
         }
